requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
feedparser==6.0.8
Jinja2==3.0.2
lxml==4.6.3
tenacity==8.0.1
Pillow==8.4.0